LLM schema conversion layer
