+++ examples/function-calling/arguments/chatgpt.example.input.json
@@ -1,4 +1,4 @@
 {
-  "name": "John Doe",
-  "age": 30
+  "name": "name",
+  "age": 0
 }
