+++ examples/function-calling/arguments/chatgpt.sale.input.json
@@ -88,7 +88,7 @@
       ],
       "stocks": [
         {
-          "name": "Surface Pro 9 i3/8GB/128GB",
+          "name": "Surface Pro 9 (i3, 8GB, 128GB)",
           "price": {
             "nominal": 1000000,
             "real": 899000
@@ -110,7 +110,7 @@
           ]
         },
         {
-          "name": "Surface Pro 9 i3/16GB/256GB",
+          "name": "Surface Pro 9 (i3, 16GB, 256GB)",
           "price": {
             "nominal": 1200000,
             "real": 1099000
@@ -132,7 +132,7 @@
           ]
         },
         {
-          "name": "Surface Pro 9 i3/16GB/512GB",
+          "name": "Surface Pro 9 (i3, 16GB, 512GB)",
           "price": {
             "nominal": 1400000,
             "real": 1299000
@@ -154,7 +154,7 @@
           ]
         },
         {
-          "name": "Surface Pro 9 i5/16GB/256GB",
+          "name": "Surface Pro 9 (i5, 16GB, 256GB)",
           "price": {
             "nominal": 1500000,
             "real": 1399000
@@ -176,7 +176,7 @@
           ]
         },
         {
-          "name": "Surface Pro 9 i5/32GB/512GB",
+          "name": "Surface Pro 9 (i5, 32GB, 512GB)",
           "price": {
             "nominal": 1800000,
             "real": 1699000
@@ -198,7 +198,7 @@
           ]
         },
         {
-          "name": "Surface Pro 9 i7/16GB/512GB",
+          "name": "Surface Pro 9 (i7, 16GB, 512GB)",
           "price": {
             "nominal": 1800000,
             "real": 1699000
@@ -220,7 +220,7 @@
           ]
         },
         {
-          "name": "Surface Pro 9 i7/32GB/512GB",
+          "name": "Surface Pro 9 (i7, 32GB, 512GB)",
           "price": {
             "nominal": 2000000,
             "real": 1899000
@@ -250,7 +250,7 @@
       "options": [],
       "stocks": [
         {
-          "name": "Extended Warranty Program",
+          "name": "Surface Pro 9 Warranty",
           "price": {
             "nominal": 100000,
             "real": 89000
@@ -267,7 +267,7 @@
       "options": [],
       "stocks": [
         {
-          "name": "Magnetic Keyboard",
+          "name": "Surface Pro 9 Keyboard",
           "price": {
             "nominal": 200000,
             "real": 169000
@@ -283,10 +283,12 @@
   ],
   "tags": [
     "Surface Pro 9",
-    "laptop",
-    "tablet",
     "2-in-1",
+    "tablet",
+    "laptop",
     "Microsoft",
-    "technology"
+    "5G",
+    "detachable keyboard",
+    "touchscreen"
   ]
 }
