+++ examples/function-calling/arguments/gemini.example.input.json
@@ -1,4 +1,4 @@
 {
   "age": 30,
-  "name": "test"
+  "name": "John"
 }
