+++ examples/function-calling/schemas/chatgpt.sale.schema.json
@@ -372,5 +372,6 @@
   "required": [
     "input"
   ],
-  "additionalProperties": false
+  "additionalProperties": false,
+  "$defs": {}
 }
