+++ test/features/llm/test_llm_schema_nullable.ts
@@ -43,7 +43,10 @@ export const test_llm_schema_union = (): void => {
   const llm: ILlmSchemaV3 | null = LlmConverterV3.schema({
     components,
     schema,
-    recursive: false,
+    config: {
+      recursive: false,
+      constraint: true,
+    },
   });
   TestValidator.equals("nullable")(llm)({
     oneOf: [
