+++ test/features/llm/v3.1/test_llm_schema_v31_ultimate_union.ts
@@ -1,14 +1,20 @@
-import { OpenApi } from "@samchon/openapi";
+import { ILlmSchemaV3_1, OpenApi } from "@samchon/openapi";
 import { LlmConverterV3_1 } from "@samchon/openapi/lib/converters/LlmConverterV3_1";
 import typia from "typia";
 
 export const test_llm_schema_v31_ultimate_union = (): void => {
-  const collection = typia.json.schemas<[IJsonSchemaCollection[]]>();
-  LlmConverterV3_1.schema({
+  const collection: IJsonSchemaCollection =
+    typia.json.schemas<[IJsonSchemaCollection]>();
+  const schema: ILlmSchemaV3_1 | null = LlmConverterV3_1.schema({
     components: collection.components,
     schema: collection.schemas[0],
-    recursive: 3,
+    $defs: {},
+    config: {
+      reference: true,
+      constraint: true,
+    },
   });
+  typia.assert<ILlmSchemaV3_1>(schema);
 };
 
 interface IJsonSchemaCollection {
